requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
redis==5.0.1
requests==2.32.3
playwright==1.46.0
